Move organisations `ShowTemplate` to shared location
This will be used by both the admin and non-admin `OrgansationsController`s
