+++ src/organisations/interfaces/show-template.interface.ts
@@ -1,5 +1,5 @@
-import { Organisation } from '../../organisation.entity';
-import { SummaryList } from '../../../common/interfaces/summary-list';
+import { Organisation } from '../organisation.entity';
+import { SummaryList } from '../../common/interfaces/summary-list';
 
 export interface ShowTemplate {
   organisation: Organisation;
